Moves AssignedCasesPage to its own file
