+++ client/app/queue/JudgeAssignTaskListView.jsx
@@ -23,6 +23,7 @@ import { LOGO_COLORS } from '../constants/AppConstants';
 import { setAttorneysOfJudge, fetchTasksAndAppealsOfAttorney, setSelectionOfTaskOfUser } from './QueueActions';
 import { sortTasks } from './utils';
 import PageRoute from '../components/PageRoute';
+import AssignedCasesPage from './AssignedCasesPage';
 
 const UnassignedCasesPage = (props) => {
   const reviewableCount = props.tasksAndAppeals.length;
@@ -41,44 +42,6 @@ const UnassignedCasesPage = (props) => {
 
   return tableContent;
 };
-
-const AssignedCasesPage = connect(
-  (state) => _.pick(state.queue, 'tasksAndAppealsOfAttorney', 'attorneysOfJudge'),
-  (dispatch) => (bindActionCreators({ setSelectionOfTaskOfUser }, dispatch)))(
-  (props) => {
-    const {
-      match, attorneysOfJudge, tasksAndAppealsOfAttorney
-    } = props;
-    const { attorneyId } = match.params;
-
-    if (!(attorneyId in tasksAndAppealsOfAttorney) || tasksAndAppealsOfAttorney[attorneyId].state === 'LOADING') {
-      return <SmallLoader message="Loading..." spinnerColor={LOGO_COLORS.QUEUE.ACCENT} />;
-    }
-
-    if (tasksAndAppealsOfAttorney[attorneyId].state === 'FAILED') {
-      const { error } = tasksAndAppealsOfAttorney[attorneyId];
-
-      return <StatusMessage title={error.response.statusText}>Error fetching cases</StatusMessage>;
-    }
-
-    const attorneyName = attorneysOfJudge.filter((attorney) => attorney.id.toString() === attorneyId)[0].full_name;
-    const { tasks, appeals } = tasksAndAppealsOfAttorney[attorneyId].data;
-
-    return <React.Fragment>
-      <h2>{attorneyName}'s Cases</h2>
-      <JudgeAssignTaskTable
-        tasksAndAppeals={
-          sortTasks({
-            tasks,
-            appeals
-          }).
-            map((task) => ({
-              task,
-              appeal: appeals[task.vacolsId] }))
-        }
-        userId={attorneyId} />
-    </React.Fragment>;
-  });
 
 class JudgeAssignTaskListView extends React.PureComponent {
   componentWillUnmount = () => {
